Add files via upload
Major changes I made was adding something for the stepsize, general debugging, switching from y to z, refining the priors, simplifying the run mcmc (because I wrote something for the big-step thing) and I also added an example to show that our rebound simulation and transit timing derivation works out quite nicely. Another thing is that I found a paper while writing the intro that gives better parameters for the masses, so we could consider those for the priors? the mcmc is just taking so much longer than expected that i didn't have the chance to add and run that. Anyway, it's the Hadden 2016 paper here! https://iopscience.iop.org/article/10.3847/0004-637X/828/1/44